+++ app/src/main/java/com/example/ghostzilla/models/coingecko/coin/Low24h.kt
@@ -5,11 +5,11 @@ import com.google.gson.annotations.SerializedName
 
 data class Low24h(
     @SerializedName("aed")
-    val aed: Int,
+    val aed: Float,
     @SerializedName("eur")
-    val eur: Int,
+    val eur: Float,
     @SerializedName("gbp")
-    val gbp: Int,
+    val gbp: Float,
     @SerializedName("usd")
-    val usd: Int
+    val usd: Float
 )
